Add templates.js to vendor + don't overwite hogan
Adding templates.js to vendor makes it possible to add `templates.js` to
the application.js sprockets file and it not complain that the file
doesn't exist in development.

Making hogan check for an instance of hogan on the page stops the
hogan template from nuking the full mustache hogan in development.