Remove blank templates.js
The presence of this file was causing issues in whitehall - it was being loaded
ahead of the local templates.js in the app. It is no longer needed as whitehall
will soon include a fully compiled version of template.js in the repository. Any
developers who modify the mustache templates will need to re-compile them and
update the repository.