Remove `typeof exports` check
Sometimes the `exports` global variable is present through a quirk of
dynamically generated heading IDs which the browser makes accessible
through the window namespace. eg <h2 id="exports">Exports</h2> becomes
`window.exports` and typeof exports is no longer undefined.

Creating a global variable for each element with an ID is a defined
HTML5 behaviour:
http://www.w3.org/html/wg/drafts/html/master/browsers.html#named-access-
on-the-window-object

The exports check can be removed altogether as we aren’t making use of
the commonjs module pattern.

This is a hotfix to a vendor asset, and should ideally be fixed
upstream.